prettified solution: separated in two modules so its more clear for the reader
